backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.0.3
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-environ==0.11.2
django-jazzmin==2.4.7
django-js-asset==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
#mysqlclient==2.1.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.3.0
pyparsing==3.1.4
python-dotenv==1.0.1
pytz==2021.3
requests==2.32.3
rsa==4.9
sqlparse==0.4.2
typing_extensions==4.12.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
